[03601] Extract PlanYamlRepairService and move RemoveWorktrees to WorktreeCleanupService
- Create PlanYamlRepairService as a static class with all YAML repair logic
  (RepairPlanYaml, NormalizePlanYamlStructure, RepairMultiLineQuotedScalars)
- Move RemoveWorktrees, ExtractSafeTitle, ClearReadOnlyAttributes from
  PlanReaderService to WorktreeCleanupService where they belong
- Split NormalizePlanYamlStructure brain method into ProcessListContext and
  ProcessStructuredListItem helper methods
- Keep PlanReaderService.RepairPlanYaml as a thin forwarder for backward compat
- Update all callers across Commands, Dialogs, Services, and Tests
