+++ src/Ivy.Tendril/Services/PlanReaderService.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics;
 using System.Globalization;
 using System.Text.RegularExpressions;
 using Ivy.Helpers;
@@ -17,7 +16,6 @@ public class PlanReaderService(
     IPlanWatcherService? planWatcherService = null) : IPlanReaderService
 {
     private static readonly Regex FolderNameRegex = new(@"^(\d{5})-(.+)$", RegexOptions.Compiled);
-    private static readonly Regex SafeTitleRegex = new(@"^\d{5}-(.+)", RegexOptions.Compiled);
     private static readonly Regex StateLineRegex = new(@"(?m)^state:\s*(.+)$", RegexOptions.Compiled);
 
     private static readonly HashSet<string> TerminalStates = new(StringComparer.OrdinalIgnoreCase)
@@ -114,7 +112,7 @@ public class PlanReaderService(
                     if (stateMatch.Success && TerminalStates.Contains(stateMatch.Groups[1].Value.Trim()))
                         continue;
 
-                    var repaired = RepairPlanYaml(yaml);
+                    var repaired = PlanYamlRepairService.RepairPlanYaml(yaml);
 
                     if (repaired != yaml)
                     {
@@ -358,7 +356,7 @@ public class PlanReaderService(
         WriteFileInBackground(() =>
         {
             if (!Directory.Exists(folderPath)) return;
-            RemoveWorktrees(folderPath, _logger, _worktreeLifecycleLogger);
+            WorktreeCleanupService.RemoveWorktrees(folderPath, _logger, _worktreeLifecycleLogger);
             WorktreeCleanupService.ForceDeleteDirectory(folderPath, _logger);
         });
     }
@@ -621,419 +619,7 @@ public class PlanReaderService(
         _useDatabaseForReads = true;
     }
 
-    /// <summary>
-    ///     Repairs common YAML issues in plan.yaml content so that YamlDotNet can parse them.
-    /// </summary>
-    internal static string RepairPlanYaml(string yaml)
-    {
-        var repaired = yaml;
-
-        // 0. Convert multi-line or unterminated quoted scalars to block scalars
-        repaired = RepairMultiLineQuotedScalars(repaired);
-
-        // 1. Remove YAML document markers (--- anywhere in the document)
-        repaired = Regex.Replace(repaired, @"(?m)^---[ \t]*(\r?\n|$)", "");
-
-        // 2. Convert object-style repos (- name: X\n    path: Y\n    branch: ...) to plain path strings
-        repaired = Regex.Replace(repaired,
-            @"(?m)^(\s*)-\s+name:\s*.+\r?\n\s+path:\s*(.+?)(?:\r?\n\s+(?:branch|prRule):\s*.+)*$",
-            "$1- $2");
-
-        // 3. Fix structured repos (- path: X + optional prRule:/branch:) → plain path strings
-        repaired = Regex.Replace(repaired,
-            @"(?m)^(\s*)-\s+path:\s*(.+?)(?:\r?\n\s+(?:prRule|branch):\s*.+)*$",
-            "$1- $2");
-
-        // 4. Convert object-style commits (- hash: X\n  repo: ...\n  message: ...) to plain hash strings
-        repaired = Regex.Replace(repaired,
-            @"(?m)^(\s*)-\s+hash:\s*(.+?)(?:\r?\n\s+(?:repo|message):\s*.+)*$",
-            "$1- $2");
-
-        // 5. Convert object-style prs (- note: X) to plain strings
-        repaired = Regex.Replace(repaired,
-            @"(?m)^(\s*)-\s+note:\s*(.+)$",
-            "$1- $2");
-
-        // 6. Remove orphan branch: lines attached to string list items (e.g. - "path"\n    branch: "")
-        repaired = Regex.Replace(repaired,
-            @"(?m)^(\s*-\s+(?:""[^""]*""|'[^']*'|%\S+).*)\r?\n\s+branch:\s*.*$",
-            "$1");
-
-        // 7. Fix double-quoted strings containing unescaped backslashes → single quotes
-        repaired = Regex.Replace(repaired, @"(?m)^([^:]+:\s+)""(.+)""(\s*)$", m =>
-        {
-            var prefix = m.Groups[1].Value;
-            var inner = m.Groups[2].Value;
-            var suffix = m.Groups[3].Value;
-            if (Regex.IsMatch(inner, @"\\[^""\\nt/abfre0 UNLP_xu]"))
-            {
-                var escaped = inner.Replace("'", "''");
-                return $"{prefix}'{escaped}'{suffix}";
-            }
-
-            return m.Value;
-        });
-
-        // 8. Fix double-quoted list items with bad escapes
-        repaired = Regex.Replace(repaired, @"(?m)^(\s*-\s+)""(.+)""(\s*)$", m =>
-        {
-            var prefix = m.Groups[1].Value;
-            var inner = m.Groups[2].Value;
-            var suffix = m.Groups[3].Value;
-            if (Regex.IsMatch(inner, @"\\[^""\\nt/abfre0 UNLP_xu]"))
-            {
-                var escaped = inner.Replace("'", "''");
-                return $"{prefix}'{escaped}'{suffix}";
-            }
-
-            return m.Value;
-        });
-
-        // 9. Quote unquoted Windows paths in list items: - D:\something → - 'D:\something'
-        repaired = Regex.Replace(repaired, @"(?m)^(\s*-\s+)([A-Za-z]:\\[^\s].*)$", m =>
-        {
-            var prefix = m.Groups[1].Value;
-            var path = m.Groups[2].Value.TrimEnd();
-            if (path.StartsWith("\"") || path.StartsWith("'")) return m.Value;
-            var escaped = path.Replace("'", "''");
-            return $"{prefix}'{escaped}'";
-        });
-
-        // 10. Quote unquoted scalar values that contain ': ' (colon-space), which YAML
-        //     misinterprets as nested mappings. Targets freeform text fields like initialPrompt.
-        repaired = Regex.Replace(repaired, @"(?m)^(\s*\w+:\s+)(.+)$", m =>
-        {
-            var prefix = m.Groups[1].Value;
-            var value = m.Groups[2].Value.TrimEnd();
-            // Skip if already quoted, is a block scalar indicator, or is a list/mapping
-            if (value.StartsWith("\"") || value.StartsWith("'") ||
-                value.StartsWith("|") || value.StartsWith(">") ||
-                value.StartsWith("-") || value.StartsWith("{") || value.StartsWith("["))
-                return m.Value;
-            // Only fix if the value contains an embedded ': ' that would confuse the parser
-            // (but not the first colon which is the key separator — we're already past it)
-            if (value.Contains(": "))
-            {
-                var escaped = value.Replace("'", "''");
-                return $"{prefix}'{escaped}'";
-            }
-
-            return m.Value;
-        });
-
-        repaired = Regex.Replace(
-            repaired,
-            @"(?m)^(\s*)(repos|commits|prs|verifications|relatedPlans|dependsOn):\s*\r?\n(?!\s*-)",
-            "$1$2: []\n");
-
-        // 12. Fix list keys with empty/whitespace-only scalar values (e.g. "relatedPlans: " at EOF)
-        //     ConvertTo-Yaml in powershell-yaml serializes empty arrays as bare "key: " lines.
-        repaired = Regex.Replace(
-            repaired,
-            @"(?m)^(\s*)(repos|commits|prs|verifications|relatedPlans|dependsOn):[ \t]*$",
-            "$1$2: []");
-
-        // 13. Fix priority field with non-numeric values (e.g. "priority: NiceToHave" should be "priority: 0")
-        repaired = Regex.Replace(
-            repaired,
-            @"(?m)^priority:\s*(?!\d).*$",
-            "priority: 0");
-
-        return NormalizePlanYamlStructure(repaired);
-    }
-
-    private static string NormalizePlanYamlStructure(string yaml)
-    {
-        var topLevelKeys = new HashSet<string>(StringComparer.Ordinal)
-        {
-            "state", "project", "level", "title", "sessionId",
-            "repos", "created", "updated", "initialPrompt", "sourceUrl",
-            "prs", "commits", "verifications", "relatedPlans", "dependsOn",
-            "priority", "executionProfile", "recommendations"
-        };
-        var listKeys = new HashSet<string>(StringComparer.Ordinal)
-        {
-            "repos", "prs", "commits", "verifications", "relatedPlans", "dependsOn",
-            "recommendations"
-        };
-
-        var normalized = yaml.Replace("\r\n", "\n");
-        var lines = normalized.Split('\n');
-        var output = new List<string>(lines.Length);
-        string? currentListKey = null;
-        var inStructuredListItem = false;
-        var inListItemBlockScalar = false;
-        var inBlockScalar = false;
-        var inUnknownKey = false;
-        var seenKeys = new HashSet<string>();
-        var structuredListKeys = new Dictionary<string, (string itemStartPattern, string subKeyPattern)>
-        {
-            ["verifications"] = (@"^-\s+name:", "^(name|status):"),
-            ["recommendations"] = (@"^-\s+title:", "^(title|description|state|impact|risk|declineReason):")
-        };
-
-        static bool IsBlockScalarValue(string value)
-        {
-            return value is "|" or "|-" or ">" or ">-";
-        }
-
-        string? TryExtractTopLevelKey(string trimmedLine, out string normalizedLine)
-        {
-            normalizedLine = trimmedLine;
-
-            var keyMatch = Regex.Match(trimmedLine, "^([A-Za-z][A-Za-z0-9]*):");
-            if (keyMatch.Success && topLevelKeys.Contains(keyMatch.Groups[1].Value))
-                return keyMatch.Groups[1].Value;
-
-            var quotedMatch = Regex.Match(trimmedLine, @"^'([A-Za-z][A-Za-z0-9]*):\s*(.*)'$");
-            if (!quotedMatch.Success || !topLevelKeys.Contains(quotedMatch.Groups[1].Value))
-                return null;
-
-            var key = quotedMatch.Groups[1].Value;
-            var value = quotedMatch.Groups[2].Value.Replace("''", "'");
-            normalizedLine = $"{key}: {value}".TrimEnd();
-            return key;
-        }
-
-        foreach (var rawLine in lines)
-        {
-            var line = rawLine.TrimEnd('\r');
-            var trimmed = line.TrimStart();
-
-            if (trimmed.Length == 0)
-            {
-                output.Add(inBlockScalar ? "  " : string.Empty);
-                continue;
-            }
-
-            var detectedKey = TryExtractTopLevelKey(trimmed, out var normalizedTopLevelLine);
-            if (inBlockScalar && detectedKey == null)
-            {
-                output.Add($"  {line}");
-                continue;
-            }
-
-            if (detectedKey != null)
-            {
-                if (!seenKeys.Add(detectedKey))
-                    continue;
-
-                var key = detectedKey;
-                currentListKey = listKeys.Contains(key) ? key : null;
-                inStructuredListItem = false;
-                inListItemBlockScalar = false;
-                inBlockScalar = false;
-                inUnknownKey = false;
-
-                if (currentListKey != null &&
-                    Regex.IsMatch(normalizedTopLevelLine, @"^[A-Za-z][A-Za-z0-9]*:\s*\[\]\s*$"))
-                    output.Add($"{key}:");
-                else
-                    output.Add(normalizedTopLevelLine);
-
-                var scalarValue = normalizedTopLevelLine[(key.Length + 1)..].Trim();
-                if (IsBlockScalarValue(scalarValue))
-                    inBlockScalar = true;
-
-                continue;
-            }
-
-            if (inUnknownKey)
-            {
-                if (line != trimmed)
-                    continue;
-                inUnknownKey = false;
-            }
-
-            if (inBlockScalar)
-            {
-                output.Add($"  {line}");
-                continue;
-            }
-
-            if (currentListKey != null)
-            {
-                var isStructured = structuredListKeys.TryGetValue(currentListKey, out var patterns);
-
-                if (trimmed.StartsWith("-"))
-                {
-                    if (isStructured && !Regex.IsMatch(trimmed, patterns.itemStartPattern))
-                    {
-                        inStructuredListItem = false;
-                        inListItemBlockScalar = false;
-                        continue;
-                    }
-
-                    output.Add($"  {trimmed}");
-                    inStructuredListItem = isStructured;
-                    inListItemBlockScalar = false;
-                }
-                else if (isStructured && inStructuredListItem)
-                {
-                    if (inListItemBlockScalar)
-                    {
-                        if (Regex.IsMatch(trimmed, patterns.subKeyPattern))
-                        {
-                            inListItemBlockScalar = false;
-                            output.Add($"    {trimmed}");
-                        }
-                        else
-                        {
-                            output.Add($"      {trimmed}");
-                        }
-                    }
-                    else if (Regex.IsMatch(trimmed, patterns.subKeyPattern))
-                    {
-                        output.Add($"    {trimmed}");
-                        var subValue = trimmed[(trimmed.IndexOf(':') + 1)..].Trim();
-                        if (IsBlockScalarValue(subValue))
-                            inListItemBlockScalar = true;
-                    }
-                }
-                else
-                {
-                    var strayKeyMatch = Regex.Match(trimmed, "^([A-Za-z][A-Za-z0-9]+):");
-                    if (strayKeyMatch.Success && topLevelKeys.Contains(strayKeyMatch.Groups[1].Value))
-                    {
-                        var key = strayKeyMatch.Groups[1].Value;
-                        currentListKey = listKeys.Contains(key) ? key : null;
-                        inStructuredListItem = false;
-                        inListItemBlockScalar = false;
-                        output.Add(trimmed);
-                    }
-                    else if (strayKeyMatch.Success)
-                    {
-                        currentListKey = null;
-                        inUnknownKey = true;
-                        inListItemBlockScalar = false;
-                    }
-                    else
-                    {
-                        output.Add($"  {trimmed}");
-                    }
-                }
-
-                continue;
-            }
-
-            var unknownKeyMatch = Regex.Match(trimmed, "^([A-Za-z][A-Za-z0-9]+):");
-            if (unknownKeyMatch.Success)
-            {
-                inUnknownKey = true;
-                continue;
-            }
-
-            output.Add(trimmed);
-        }
-
-        return string.Join(Environment.NewLine, output);
-    }
-
-    private static string RepairMultiLineQuotedScalars(string yaml)
-    {
-        var normalized = yaml.Replace("\r\n", "\n");
-        var lines = normalized.Split('\n');
-        var result = new List<string>();
-
-        for (var i = 0; i < lines.Length; i++)
-        {
-            var m = Regex.Match(lines[i], @"^([a-zA-Z]\w*:)\s+([""'])(.*)$");
-            if (!m.Success)
-            {
-                result.Add(lines[i]);
-                continue;
-            }
-
-            var key = m.Groups[1].Value;
-            var quote = m.Groups[2].Value[0];
-            var rest = m.Groups[3].Value;
-
-            if (IsScalarQuoteClosed(rest, quote))
-            {
-                result.Add(lines[i]);
-                continue;
-            }
-
-            var contentLines = new List<string>();
-            if (rest.Length > 0)
-                contentLines.Add(rest);
-
-            var j = i + 1;
-            while (j < lines.Length)
-            {
-                var next = lines[j];
-                var nextTrimmed = next.TrimStart();
-
-                if (nextTrimmed.Length > 0 && Regex.IsMatch(nextTrimmed, @"^[a-zA-Z]\w*:(\s|$)"))
-                    break;
-
-                var nextStripped = next.TrimEnd();
-                if (nextStripped.Length > 0 && EndsWithClosingQuote(nextStripped, quote))
-                {
-                    contentLines.Add(nextStripped[..^1]);
-                    j++;
-                    break;
-                }
-
-                contentLines.Add(next);
-                j++;
-            }
-
-            while (contentLines.Count > 0 && string.IsNullOrWhiteSpace(contentLines[^1]))
-                contentLines.RemoveAt(contentLines.Count - 1);
-
-            result.Add($"{key} |");
-            foreach (var cl in contentLines)
-            {
-                var unescaped = quote == '"'
-                    ? cl.Replace("\\\\", "\x01").Replace("\\\"", "\"").Replace("\x01", "\\")
-                    : cl.Replace("''", "'");
-                result.Add($"  {unescaped}");
-            }
-
-            i = j - 1;
-        }
-
-        return string.Join("\n", result);
-    }
-
-    private static bool IsScalarQuoteClosed(string afterOpenQuote, char quote)
-    {
-        if (quote == '"')
-        {
-            for (var k = 0; k < afterOpenQuote.Length; k++)
-            {
-                if (afterOpenQuote[k] == '\\') { k++; continue; }
-                if (afterOpenQuote[k] == '"') return true;
-            }
-            return false;
-        }
-
-        for (var k = 0; k < afterOpenQuote.Length; k++)
-        {
-            if (afterOpenQuote[k] != '\'') continue;
-            if (k + 1 < afterOpenQuote.Length && afterOpenQuote[k + 1] == '\'') { k++; continue; }
-            return true;
-        }
-        return false;
-    }
-
-    private static bool EndsWithClosingQuote(string line, char quote)
-    {
-        if (line.Length == 0 || line[^1] != quote) return false;
-        if (quote == '"')
-        {
-            var backslashes = 0;
-            for (var k = line.Length - 2; k >= 0 && line[k] == '\\'; k--)
-                backslashes++;
-            return backslashes % 2 == 0;
-        }
-
-        var quotes = 0;
-        for (var k = line.Length - 1; k >= 0 && line[k] == '\''; k--)
-            quotes++;
-        return quotes % 2 == 1;
-    }
+    internal static string RepairPlanYaml(string yaml) => PlanYamlRepairService.RepairPlanYaml(yaml);
 
     /// <summary>
     ///     Enumerates all directories in PlansDirectory that match the plan folder
@@ -1118,124 +704,8 @@ public class PlanReaderService(
         return latestFile != null ? FileHelper.ReadAllText(latestFile) : string.Empty;
     }
 
-    internal static void RemoveWorktrees(string planFolderPath, ILogger? logger = null, IWorktreeLifecycleLogger? lifecycleLogger = null)
-    {
-        var worktreesDir = Path.Combine(planFolderPath, "worktrees");
-        if (!Directory.Exists(worktreesDir)) return;
 
-        var planId = WorktreeLifecycleLogger.ExtractPlanId(planFolderPath);
 
-        // Extract safe title from plan folder name for branch deletion
-        var safeTitle = ExtractSafeTitle(planFolderPath);
-        var branchName = $"tendril/{planId}-{safeTitle}";
-
-        foreach (var wtDir in Directory.GetDirectories(worktreesDir))
-        {
-            var gitFile = Path.Combine(wtDir, ".git");
-            if (!File.Exists(gitFile))
-            {
-                var dirAge = DateTime.UtcNow - new DirectoryInfo(wtDir).CreationTimeUtc;
-                logger?.LogInformation(
-                    "Worktree directory has no .git file (created {Age} ago), force-deleting: {Path}",
-                    dirAge, Path.GetFileName(wtDir));
-                lifecycleLogger?.LogCleanupAttempt(planId, wtDir, "RemoveWorktrees(force)", gitFileExists: false);
-
-                try
-                {
-                    WorktreeCleanupService.ForceDeleteDirectory(wtDir, logger);
-                    lifecycleLogger?.LogCleanupSuccess(planId, wtDir);
-                }
-                catch (Exception ex)
-                {
-                    lifecycleLogger?.LogCleanupFailed(planId, wtDir, ex.Message);
-                    logger?.LogWarning(ex, "Failed to force-delete worktree directory {Dir}", Path.GetFileName(wtDir));
-                }
-                continue;
-            }
-
-            // Read the .git file to find which repo this worktree belongs to.
-            // Format: "gitdir: <path-to-repo>/.git/worktrees/<name>"
-            var gitContent = FileHelper.ReadAllText(gitFile).Trim();
-            var match = Regex.Match(gitContent, @"gitdir:\s*(.+)");
-            if (!match.Success) continue;
-
-            var gitDir = match.Groups[1].Value.Trim();
-            // Navigate from .git/worktrees/<name> up to the repo root
-            var repoGitDir = Path.GetFullPath(Path.Combine(gitDir, "..", ".."));
-            var repoRoot = Path.GetDirectoryName(repoGitDir);
-            if (repoRoot == null || !Directory.Exists(repoRoot)) continue;
-
-            lifecycleLogger?.LogCleanupAttempt(planId, wtDir, "RemoveWorktrees", gitFileExists: true);
-
-            try
-            {
-                var psi = new ProcessStartInfo("git", $"worktree remove --force \"{wtDir}\"")
-                {
-                    WorkingDirectory = repoRoot,
-                    RedirectStandardOutput = true,
-                    RedirectStandardError = true,
-                    UseShellExecute = false,
-                    CreateNoWindow = true
-                };
-                using var process = Process.Start(psi);
-                process.WaitForExitOrKill(10000);
-                lifecycleLogger?.LogCleanupSuccess(planId, wtDir);
-
-                // Delete the associated git branch
-                try
-                {
-                    var branchPsi = new ProcessStartInfo("git", $"branch -D \"{branchName}\"")
-                    {
-                        WorkingDirectory = repoRoot,
-                        RedirectStandardOutput = true,
-                        RedirectStandardError = true,
-                        UseShellExecute = false,
-                        CreateNoWindow = true
-                    };
-                    using var branchProcess = Process.Start(branchPsi);
-                    branchProcess.WaitForExitOrKill(10000);
-                    logger?.LogInformation("Deleted branch {BranchName} for worktree {WorktreeDir}", branchName, Path.GetFileName(wtDir));
-                }
-                catch (Exception branchEx)
-                {
-                    logger?.LogWarning(branchEx, "Failed to delete branch {BranchName} for worktree {WorktreeDir}", branchName, Path.GetFileName(wtDir));
-                }
-            }
-            catch (Exception ex)
-            {
-                lifecycleLogger?.LogCleanupFailed(planId, wtDir, ex.Message);
-            }
-        }
-    }
-
-    /// <summary>
-    ///     Extracts the safe title from a plan folder path using regex pattern matching.
-    /// </summary>
-    internal static string ExtractSafeTitle(string planFolderPath)
-    {
-        if (string.IsNullOrEmpty(planFolderPath))
-            return "Unknown";
-        var folderName = Path.GetFileName(planFolderPath.TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar));
-        var match = SafeTitleRegex.Match(folderName);
-        return match.Success ? match.Groups[1].Value : "Unknown";
-    }
-
-    internal static void ClearReadOnlyAttributes(string directoryPath)
-    {
-        try
-        {
-            foreach (var file in Directory.EnumerateFiles(directoryPath, "*", SearchOption.AllDirectories))
-            {
-                var attrs = File.GetAttributes(file);
-                if ((attrs & FileAttributes.ReadOnly) != 0)
-                    File.SetAttributes(file, attrs & ~FileAttributes.ReadOnly);
-            }
-        }
-        catch
-        {
-            // Best-effort
-        }
-    }
 
     /// <summary>
     ///     Parses a single plan folder from the file system. Used by the sync service for incremental updates.
@@ -1268,7 +738,7 @@ public class PlanReaderService(
             {
                 // Fall back to the repair pass for malformed agent-generated YAML.
                 _logger.LogWarning(ex, "Failed to parse plan YAML {PlanYamlPath}, attempting repair", planYamlPath);
-                var repaired = RepairPlanYaml(yamlContent);
+                var repaired = PlanYamlRepairService.RepairPlanYaml(yamlContent);
                 if (repaired != yamlContent)
                     FileHelper.WriteAllText(planYamlPath, repaired);
